Specify java version for PME also
