+++ repour/adjust/gradle_provider.py
@@ -17,8 +17,6 @@ EXECUTION_NAME = "GRADLE"
 INIT_SCRIPT_FILE_NAME = "analyzer-init.gradle"
 MANIPULATION_FILE_NAME = "manipulation.json"
 
-REPOUR_JAVA_KEY = "-DRepour_Java="
-
 def get_gradle_provider(init_file_path, default_parameters, specific_indy_group=None, timestamp=None):
 
     async def adjust(work_dir, extra_adjust_parameters, adjust_result):
@@ -58,7 +56,7 @@ def get_gradle_provider(init_file_path, default_parameters, specific_indy_group=
             live_log=True
         )
 
-        jvm_version = get_jvm_from_extra_parameters(extra_parameters)
+        jvm_version = util.get_jvm_from_extra_parameters(extra_parameters)
 
         if jvm_version:
             env = {'JAVA_HOME': '/usr/lib/jvm/java-' + jvm_version + '-openjdk'}
@@ -146,16 +144,3 @@ def get_command_gradle(work_dir):
         command_gradle = './gradlew'
 
     return command_gradle
-
-
-def get_jvm_from_extra_parameters(extra_parameters):
-    """
-    If repour JVM option specified, return the option value. Otherwise return None
-    """
-
-    for parameter in extra_parameters:
-
-        if REPOUR_JAVA_KEY in parameter:
-            return parameter.replace(REPOUR_JAVA_KEY, '')
-    else:
-        return None
